+++ src/main/kotlin/com/petqua/application/product/ProductService.kt
@@ -1,11 +1,11 @@
-package com.petqua.application
+package com.petqua.application.product
 
+import com.petqua.application.product.dto.ProductDetailResponse
+import com.petqua.application.product.dto.ProductReadRequest
+import com.petqua.application.product.dto.ProductsResponse
 import com.petqua.common.domain.findByIdOrThrow
-import com.petqua.domain.ProductRepository
-import com.petqua.domain.StoreRepository
-import com.petqua.dto.ProductDetailResponse
-import com.petqua.dto.ProductReadRequest
-import com.petqua.dto.ProductsResponse
+import com.petqua.domain.product.ProductRepository
+import com.petqua.domain.store.StoreRepository
 import com.petqua.exception.product.ProductException
 import com.petqua.exception.product.ProductExceptionType.NOT_FOUND_PRODUCT
 import com.petqua.exception.store.StoreException
