+++ src/main/kotlin/com/petqua/domain/product/ProductCustomRepository.kt
@@ -1,8 +1,8 @@
-package com.petqua.domain
+package com.petqua.domain.product
 
-import com.petqua.dto.ProductPaging
-import com.petqua.dto.ProductReadCondition
-import com.petqua.dto.ProductResponse
+import com.petqua.domain.product.dto.ProductPaging
+import com.petqua.domain.product.dto.ProductReadCondition
+import com.petqua.domain.product.dto.ProductResponse
 
 interface ProductCustomRepository {
