+++ src/main/kotlin/com/petqua/domain/product/ProductSourceType.kt
@@ -1,4 +1,4 @@
-package com.petqua.domain
+package com.petqua.domain.product
 
 enum class ProductSourceType {
     NONE, HOME_RECOMMENDED, HOME_NEW_ENROLLMENT
