+++ src/main/kotlin/com/petqua/domain/product/Sorter.kt
@@ -1,4 +1,4 @@
-package com.petqua.domain
+package com.petqua.domain.product
 
 enum class Sorter {
     NONE, SALE_PRICE_ASC, SALE_PRICE_DESC, REVIEW_COUNT_DESC, ENROLLMENT_DATE_DESC
