+++ src/main/kotlin/com/petqua/domain/recommendation/ProductRecommendationRepository.kt
@@ -1,4 +1,4 @@
-package com.petqua.domain
+package com.petqua.domain.recommendation
 
 import org.springframework.data.jpa.repository.JpaRepository
