+++ src/main/kotlin/com/petqua/domain/store/StoreRepository.kt
@@ -1,4 +1,4 @@
-package com.petqua.domain
+package com.petqua.domain.store
 
 import org.springframework.data.jpa.repository.JpaRepository
