+++ src/test/kotlin/com/petqua/domain/ProductCustomRepositoryImplTest.kt
@@ -1,13 +1,16 @@
 package com.petqua.domain
 
-import com.petqua.domain.ProductSourceType.HOME_RECOMMENDED
-import com.petqua.domain.Sorter.ENROLLMENT_DATE_DESC
-import com.petqua.domain.Sorter.REVIEW_COUNT_DESC
-import com.petqua.domain.Sorter.SALE_PRICE_ASC
-import com.petqua.domain.Sorter.SALE_PRICE_DESC
-import com.petqua.dto.ProductPaging
-import com.petqua.dto.ProductReadCondition
-import com.petqua.dto.ProductResponse
+import com.petqua.domain.product.ProductRepository
+import com.petqua.domain.product.ProductSourceType.HOME_RECOMMENDED
+import com.petqua.domain.product.Sorter.ENROLLMENT_DATE_DESC
+import com.petqua.domain.product.Sorter.REVIEW_COUNT_DESC
+import com.petqua.domain.product.Sorter.SALE_PRICE_ASC
+import com.petqua.domain.product.Sorter.SALE_PRICE_DESC
+import com.petqua.domain.product.dto.ProductPaging
+import com.petqua.domain.product.dto.ProductReadCondition
+import com.petqua.domain.product.dto.ProductResponse
+import com.petqua.domain.recommendation.ProductRecommendationRepository
+import com.petqua.domain.store.StoreRepository
 import com.petqua.test.DataCleaner
 import com.petqua.test.fixture.product
 import com.petqua.test.fixture.productRecommendation
