+++ src/test/kotlin/com/petqua/test/fixture/StoreFixtures.kt
@@ -1,6 +1,6 @@
 package com.petqua.test.fixture
 
-import com.petqua.domain.Store
+import com.petqua.domain.store.Store
 
 fun store(
     id: Long = 0L,
